+++ src/Main.cpp
@@ -203,6 +203,39 @@ void SS2K::maintenanceLoop(void *pvParameters) {
     // If we're in ERG mode, modify shift commands to inc/dec the target watts instead.
     ss2k->FTMSModeShiftModifier();
 
+    // If we have a resistance bike attached, slow down when we're close to the limits.
+    if (rtConfig->resistance.getValue() > 0) {
+      int speed           = userConfig->getStepperSpeed();
+      float resistance    = rtConfig->resistance.getValue();
+      float maxResistance = rtConfig->getMaxResistance();
+
+      // Slow down when resistance is within 20% of the lower limit
+      if (resistance < (maxResistance * 0.2)) {
+        float factor = resistance / (maxResistance * 0.2);
+        speed        = static_cast<int>(factor * userConfig->getStepperSpeed());
+        if (speed < 500) {
+          speed = 500;
+        }
+        if (ss2k->targetPosition < stepper->getCurrentPosition()) {
+          speed = userConfig->getStepperSpeed();
+        }
+      }
+
+      // Slow down when resistance is within 20% of the upper limit
+      if (resistance > (maxResistance * 0.8)) {
+        float factor = (maxResistance - resistance) / (maxResistance * 0.2);
+        speed        = static_cast<int>(factor * userConfig->getStepperSpeed());
+        if (speed < 500) {
+          speed = 500;
+        }
+        if (ss2k->targetPosition < stepper->getCurrentPosition()) {
+          speed = userConfig->getStepperSpeed();
+        }
+      }
+
+      ss2k->updateStepperSpeed(speed);
+    }
+
     // if this hardware version has serial pins, check and process their data.
     if (currentBoard.auxSerialTxPin) {
       ss2k->txSerial();
@@ -275,9 +308,9 @@ void SS2K::maintenanceLoop(void *pvParameters) {
 
     // Things to do every 20 loops
     if (loopCounter > 20) {
-     // Removed driver temp checking. This really doesn't do anything benificial anyway, because the thermistors in the ESP32 are not accurate.
-     // The Driver itsself will throttle automatically if the temp is too high. 
-     // ss2k->checkDriverTemperature();
+      // Removed driver temp checking. This really doesn't do anything benificial anyway, because the thermistors in the ESP32 are not accurate.
+      // The Driver itsself will throttle automatically if the temp is too high.
+      // ss2k->checkDriverTemperature();
 
 #ifdef DEBUG_STACK
       Serial.printf("Step Task: %d \n", uxTaskGetStackHighWaterMark(moveStepperTask));
@@ -344,9 +377,9 @@ void SS2K::FTMSModeShiftModifier() {
             ((ss2k->targetPosition + shiftDelta * userConfig->getShiftStep()) > rtConfig->getMaxStep())) {
           SS2K_LOG(MAIN_LOG_TAG, "Shift Blocked by stepper limits.");
           rtConfig->setShifterPosition(ss2k->lastShifterPosition);
-        } else if ((rtConfig->resistance.getValue() < rtConfig->getMinResistance()) && (shiftDelta > 0)) {
+        } else if ((rtConfig->resistance.getValue() <= rtConfig->getMinResistance()) && (shiftDelta > 0)) {
           // User Shifted in the proper direction - allow
-        } else if ((rtConfig->resistance.getValue() > rtConfig->getMaxResistance()) && (shiftDelta < 0)) {
+        } else if ((rtConfig->resistance.getValue() >= rtConfig->getMaxResistance()) && (shiftDelta < 0)) {
           // User Shifted in the proper direction - allow
         } else if ((rtConfig->resistance.getValue() > rtConfig->getMinResistance()) && (rtConfig->resistance.getValue() < rtConfig->getMaxResistance())) {
           // User Shifted in bounds - allow
@@ -402,9 +435,17 @@ void SS2K::moveStepper(void *pvParameters) {
         if ((rtConfig->resistance.getValue() >= rtConfig->getMinResistance()) && (rtConfig->resistance.getValue() <= rtConfig->getMaxResistance())) {
           stepper->moveTo(ss2k->targetPosition);
         } else if (rtConfig->resistance.getValue() < rtConfig->getMinResistance()) {  // Limit Stepper to Min Resistance
-          stepper->moveTo(stepper->getCurrentPosition() + 10);
+          stepper->moveTo(stepper->getCurrentPosition() + 20);
+          // Let the user Shift Out of this Position
+          if (ss2k->targetPosition > stepper->getCurrentPosition()) {
+            stepper->moveTo(ss2k->targetPosition);
+          }
         } else {  // Limit Stepper to Max Resistance
-          stepper->moveTo(stepper->getCurrentPosition() - 10);
+          stepper->moveTo(stepper->getCurrentPosition() - 20);
+          // Let the user Shift Out of this Position
+          if (ss2k->targetPosition < stepper->getCurrentPosition()) {
+            stepper->moveTo(ss2k->targetPosition);
+          }
         }
 
       } else {
@@ -531,11 +572,13 @@ void SS2K::updateStealthChop() {
   SS2K_LOG(MAIN_LOG_TAG, "StealthChop is now %d", t_bool);
 }
 
-// Applies current stepper speed
-void SS2K::updateStepperSpeed() {
-  int t = userConfig->getStepperSpeed();
-  stepper->setSpeedInHz(t);
-  SS2K_LOG(MAIN_LOG_TAG, "StepperSpeed is now %d", t);
+// Applies userconfig stepper speed if speed not specified
+void SS2K::updateStepperSpeed(int speed) {
+  if (speed == 0) {
+    speed = userConfig->getStepperSpeed();
+    SS2K_LOG(MAIN_LOG_TAG, "StepperSpeed is now %d", speed);
+  }
+  stepper->setSpeedInHz(speed);
 }
 
 // Checks the driver temperature and throttles power if above threshold.